iOS-6: Фикс коммов
